feat: add shared clipboard helper
