+++ src/lib/clipboard.ts
@@ -1,0 +1,16 @@
+// src/lib/clipboard.ts
+// Shared clipboard helper with textarea fallback.
+
+export async function copyText(text: string): Promise<void> {
+  try {
+    await navigator.clipboard.writeText(text);
+  } catch {
+    const ta = document.createElement("textarea");
+    ta.value = text;
+    document.body.appendChild(ta);
+    ta.select();
+    document.execCommand("copy");
+    ta.remove();
+  }
+}
+
